+++ samples/SettingsSample/SettingsSample/Helpers/Settings.cs
@@ -1,43 +1,25 @@
-// Helpers/Settings.cs
-using Refractored.Xam.Settings;
-using Refractored.Xam.Settings.Abstractions;
+﻿// Helpers/Settings.cs
+using Plugin.Settings;
+using Plugin.Settings.Abstractions;
 
 namespace SettingsSample.Helpers
 {
-  /// <summary>
-  /// This is the Settings static class that can be used in your Core solution or in any
-  /// of your client applications. All settings are laid out the same exact way with getters
-  /// and setters. 
-  /// </summary>
-  public static class Settings
-  {
-    private static ISettings AppSettings
-    {
-      get
-      {
-        return CrossSettings.Current;
-      }
-    }
+	/// <summary>
+	/// This is the Settings static class that can be used in your Core solution or in any
+	/// of your client applications. All settings are laid out the same exact way with getters
+	/// and setters. 
+	/// </summary>
+	public static class Settings
+	{
+		private static ISettings AppSettings => CrossSettings.Current;
 
-    #region Setting Constants
+		public static string GeneralSettings
+		{
+			get => AppSettings.GetValueOrDefault(nameof(GeneralSettings), string.Empty);
+			
+			set => AppSettings.AddOrUpdateValue(nameof(GeneralSettings), value);
+			
+		}
 
-    private const string SettingsKey = "settings_key";
-    private static readonly string SettingsDefault = string.Empty;
-
-    #endregion
-
-
-    public static string GeneralSettings
-    {
-      get
-      {
-        return AppSettings.GetValueOrDefault(SettingsKey, SettingsDefault);
-      }
-      set
-      {
-        AppSettings.AddOrUpdateValue(SettingsKey, value);
-      }
-    }
-
-  }
+	}
 }