+++ CHANGELOG.md
@@ -1,6 +1,15 @@
-### 2.6.0-beta
-* Ability to store a setting in different file/container/usersettings (simply pass file name in as last parameter)
-* Fixed issues with double parsing on Android
+### 3.0.0
+* Strongly Type all Gets and Sets
+* Re-targeted to .NET Standard 1.0
+* - Drop support for Windows Phone 8 Silverlight/Windows Phone/Store RT 8.1
+
+* Added ability to specify file name
+* -iOS: SuiteName
+* -Android: SharedPreferences name
+* -UWP/WinRT Container Name
+
+### 2.5.8
+* Fixed possible null reference exception when type is not supported.
 
 ### 2.5.4
 * Add macOS Support
@@ -12,6 +21,7 @@
 ### 2.1.0
 * Remove help file and add in readme.txt
 * .NET 4.5 Implementation for unit testing
+
 ### 2.0.0
 * Breaking changes:
 * -New namespace - Plugin.Settings